+++ lmu/__init__.py
@@ -1,6 +1,14 @@
 """LMU provides a package for deep learning with Legendre Memory Units."""
 
-from .lmu import Legendre, InputScaled, LMUCell, LMUCellODE, LMUCellGating
+from .lmu import (
+    Legendre,
+    InputScaled,
+    LMUCell,
+    LMUCellODE,
+    LMUCellGating,
+    LMUCellFFT,
+    LMU,
+)
 
 from .version import version as __version__